Move ui_move_cursor to ui.c from api/ui.c
Functions in api/ui.c require to begin with "nvim_" but the
said function is not an nvim function, hence we move it to ui.c

This involves setting the ext_win_channel in ui.c so that it knows
which channel to send RPC call to. This is done using
ui_ext_win_set_channel().
