+++ src/nvim/ui.c
@@ -36,6 +36,7 @@
 #include "nvim/ui_compositor.h"
 #include "nvim/window.h"
 #include "nvim/cursor_shape.h"
+#include "nvim/msgpack_rpc/channel.h"
 #ifdef FEAT_TUI
 # include "nvim/tui/tui.h"
 #else
@@ -59,6 +60,7 @@ static int busy = 0;
 static bool pending_mode_info_update = false;
 static bool pending_mode_update = false;
 static handle_T cursor_grid_handle = DEFAULT_GRID_HANDLE;
+static uint64_t ext_win_ui_channel = 0;
 
 #if MIN_LOG_LEVEL > DEBUG_LOG_LEVEL
 # define UI_LOG(funname)
@@ -496,5 +498,44 @@ void ui_grid_resize(handle_T grid_handle, int width, int height, Error *error)
     wp->w_height_request = (int)MAX(height, 0);
     wp->w_width_request = (int)MAX(width, 0);
     win_set_inner_size(wp);
+  }
+}
+
+void ui_set_ext_win_channel(uint64_t channel_id)
+{
+  ext_win_ui_channel = channel_id;
+}
+
+Integer ui_win_move_cursor(Integer direction, Integer count)
+{
+  Array args = ARRAY_DICT_INIT;
+  Error error = ERROR_INIT;
+  typval_T rettv;
+
+  ADD(args, INTEGER_OBJ(direction));
+  ADD(args, INTEGER_OBJ(count));
+
+  Object result = rpc_send_call(ext_win_ui_channel, "win_move_cursor", args,
+                                &error);
+
+  if (ERROR_SET(&error)) {
+    EMSG2(_("Error invoking win_move_cursor: %s"), error.msg);
+    api_clear_error(&error);
+    api_free_object(result);
+    return 0;
+  }
+
+  if (result.type != kObjectTypeInteger) {
+    api_set_error(&error, kErrorTypeValidation,
+                  "Error converting the call result: %s", error.msg);
+    api_clear_error(&error);
+    api_free_object(result);
+    return 0;
+  }
+
+  if (rettv.v_type == VAR_NUMBER) {
+    return (Integer)rettv.vval.v_number;
+  } else {
+    return 0;
   }
 }
